+++ simtong-presentation/src/main/kotlin/team/comit/simtong/domain/common/dto/FindEmployeeNumberWebRequest.kt
@@ -1,4 +1,4 @@
-package team.comit.simtong.domain.common.dto.request
+package team.comit.simtong.domain.common.dto
 
 import java.util.UUID
 import javax.validation.constraints.Email
