+++ simtong-presentation/src/main/kotlin/team/comit/simtong/domain/email/dto/SendAuthCodeWebRequest.kt
@@ -1,4 +1,4 @@
-package team.comit.simtong.domain.email.dto.request
+package team.comit.simtong.domain.email.dto
 
 import javax.validation.constraints.Email
 import javax.validation.constraints.NotBlank
