+++ simtong-presentation/src/main/kotlin/team/comit/simtong/domain/holiday/dto/AppointHolidayPeriodWebRequest.kt
@@ -1,4 +1,4 @@
-package team.comit.simtong.domain.holiday.dto.request
+package team.comit.simtong.domain.holiday.dto
 
 import org.hibernate.validator.constraints.Range
 import java.time.LocalDate
