+++ simtong-presentation/src/main/kotlin/team/comit/simtong/domain/menu/dto/SaveMenuWebRequest.kt
@@ -1,4 +1,4 @@
-package team.comit.simtong.domain.menu.dto.request
+package team.comit.simtong.domain.menu.dto
 
 import org.springframework.web.multipart.MultipartFile
 import javax.validation.constraints.NotNull
