+++ simtong-presentation/src/main/kotlin/team/comit/simtong/domain/schedule/dto/AddIndividualScheduleWebRequest.kt
@@ -1,20 +1,20 @@
-package team.comit.simtong.domain.schedule.dto.request
+package team.comit.simtong.domain.schedule.dto
 
 import org.hibernate.validator.constraints.Length
 import java.time.LocalDate
-import javax.validation.constraints.FutureOrPresent
+import java.time.LocalTime
 import javax.validation.constraints.NotBlank
 import javax.validation.constraints.NotNull
 
 /**
  *
- * 관리자가 지점 일정을 추가 요청하는 AddSpotScheduleWebRequest
+ * 개인 일정 추가를 요청하는 AddIndividualScheduleWebRequest
  *
  * @author Chokyunghyeon
- * @date 2022/11/21
+ * @date 2022/11/26
  * @version 1.0.0
  **/
-data class AddSpotScheduleWebRequest(
+data class AddIndividualScheduleWebRequest(
 
     @field:NotBlank
     @field:Length(max = 20)
@@ -25,5 +25,7 @@ data class AddSpotScheduleWebRequest(
 
     // TODO 시작일, 종료일 검증 Resolver 구현
     @field:NotNull
-    val endAt: LocalDate
+    val endAt: LocalDate,
+
+    val alarm: LocalTime?
 )
